+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 ### Added
+- Initial implementation of `SSwitch` component.
 - Initial implementation of `SRadioGroup` component.
 - `SRadioGroupItem` component to be used inside `SRadioGroup`.
 - Initial implementation of `SFilter` component.